Move callback (handle) code back into bot.go
It's little code and really belongs in bot.go anyway
